requirements.txt
bleak
asyncio
pycryptodome
